Upgrade to Tauri 2.0
