+++ package.json
@@ -16,21 +16,26 @@
 	"dependencies": {
 		"@fontsource/roboto-mono": "^5.0.16",
 		"@mdi/js": "^7.4.47",
-		"@tauri-apps/api": "^1.5.3",
+		"@tauri-apps/api": "next",
+		"@tauri-apps/plugin-dialog": "^2.0.0-alpha.5",
+		"@tauri-apps/plugin-fs": "^2.0.0-alpha.6",
+		"@tauri-apps/plugin-process": "2.0.0-alpha.5",
+		"@tauri-apps/plugin-shell": "^2.0.0-alpha.5",
+		"@tauri-apps/plugin-updater": "2.0.0-alpha.5",
+		"@tauri-apps/plugin-log": "2.0.0-alpha.5",
+		"@tauri-apps/plugin-store": "2.0.0-alpha.5",
 		"dompurify": "^3.0.8",
 		"marked": "^11.1.1",
 		"mitt": "^3.0.1",
 		"pinia": "^2.1.7",
 		"semver": "^7.5.4",
-		"tauri-plugin-log-api": "github:tauri-apps/tauri-plugin-log#v1",
-		"tauri-plugin-store-api": "github:tauri-apps/tauri-plugin-store#v1",
 		"vue": "^3.4.5",
 		"vue-router": "^4.2.5",
 		"vuetify": "^3.4.9"
 	},
 	"devDependencies": {
 		"@eslint/eslintrc": "^3.0.0",
-		"@tauri-apps/cli": "^1.5.9",
+		"@tauri-apps/cli": "next",
 		"@vitejs/plugin-vue": "^5.0.2",
 		"eslint": "^8.56.0",
 		"eslint-config-prettier": "^9.1.0",
