+++ ui/src/components/SetupGuideDialog.vue
@@ -186,7 +186,7 @@
 
 <script setup>
 import { ref, computed, onMounted } from 'vue';
-import { info } from 'tauri-plugin-log-api';
+import { info } from '@tauri-apps/plugin-log';
 import { mdiDownload } from '@mdi/js';
 
 import useSettings from '../composables/settings';
