+++ ui/src/composables/app.js
@@ -1,5 +1,5 @@
 import { ref } from 'vue';
-import { invoke } from '@tauri-apps/api';
+import { invoke } from '@tauri-apps/api/core';
 
 let name = ref('');
 let identifier = ref('');
